Crate async_github

Ratatui Async example

This example demonstrates how to use Ratatui with widgets that fetch data asynchronously. It uses the octocrab crate to fetch a list of pull requests from the GitHub API.

https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/managing-your-personal-access-tokens#creating-a-fine-grained-personal-access-token https://github.com/settings/tokens/new to create a new token (select classic, and no scopes)

This example does not cover message passing between threads, it only demonstrates how to manage shared state between the main thread and a background task, which acts mostly as a one-shot fetcher. For more complex scenarios, you may need to use channels or other synchronization primitives.

A simple app might have multiple widgets that fetch data from different sources, and each widget would have its own background task to fetch the data. The main thread would then render the widgets with the latest data.

The latest version of this example is available in the examples folder in the repository.

Please note that the examples are designed to be run against the main branch of the Github repository. This means that you may not be able to compile with the latest release version on crates.io, or the one that you have installed locally.

See the examples readme for more information on finding examples that match the version of the library you are using.

Structs

App 🔒

PullRequest 🔒

PullRequestListState 🔒

PullRequestListWidget 🔒

A widget that displays a list of pull requests.

Enums

LoadingState 🔒

Functions

main 🔒

Type Aliases

OctoPullRequest 🔒